+++ src/client/components/view_global_pane/view_global_pane.js
@@ -10,9 +10,22 @@
  *  - pr_prefix: the prefix of the checkbox nodes
  */
 
+import { pwixI18n } from 'meteor/pwix:i18n';
+
 import './view_global_pane.html';
 
 Template.view_global_pane.helpers({
+
+    // whether we have some global roles to show
+    haveGlobal( arg ){
+        return this.roles.get().global.direct.length > 0;
+    },
+
+    // string translation
+    i18n( arg ){
+        return pwixI18n.label( I18N, arg.hash.key );
+    },
+
     // parms to the pr_tree component to edit global roles
     parmsTree(){
         return {